TVIST1-425: Refactor CaseReminderController to contain index route
